+++ confirmation/models.py
@@ -123,6 +123,15 @@ _properties = {
     Confirmation.REALM_CREATION: ConfirmationType('check_prereg_key_and_redirect'),
 }
 
+def one_click_unsubscribe_link(user_profile: UserProfile, email_type: str) -> str:
+    """
+    Generate a unique link that a logged-out user can visit to unsubscribe from
+    Zulip e-mails without having to first log in.
+    """
+    return create_confirmation_link(user_profile, user_profile.realm.host,
+                                    Confirmation.UNSUBSCRIBE,
+                                    url_args = {'email_type': email_type})
+
 # Functions related to links generated by the generate_realm_creation_link.py
 # management command.
 # Note that being validated here will just allow the user to access the create_realm
