+++ zerver/lib/digest.py
@@ -10,7 +10,8 @@ from django.template import loader
 from django.conf import settings
 from django.utils.timezone import now as timezone_now
 
-from zerver.lib.notifications import build_message_list, one_click_unsubscribe_link
+from confirmation.models import one_click_unsubscribe_link
+from zerver.lib.notifications import build_message_list
 from zerver.lib.send_email import send_future_email, FromAddress
 from zerver.lib.url_encoding import encode_stream
 from zerver.models import UserProfile, UserMessage, Recipient, Stream, \
